requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==2.7.0
torch>=2.2.0
scikit-learn==1.4.2
requests==2.32.3
beautifulsoup4==4.12.3
